+++ src/components/shared/Button.tsx
@@ -8,8 +8,8 @@ interface Props {
 }
 
 const baseStyle: string = `
-block w-full sm:w-max text-center py-2 px-4 border-solid border-2 rounded transition-all hover:scale-110
-bg-white text-black border-white`;
+block w-full sm:w-max text-center py-2 px-4 border-solid border-2 rounded transition-all hover:scale-105
+bg-white text-black border-white flex justify-center`;
 
 function Button({children, url, className}: Props) {
   return (
